# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"Product": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"vendor": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"description",
					"vendor",
					"image",
					"price",
					"categories"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "senior-backend-developer-node",
		"version": "1.0.0",
		"description": "Senior Developer Node.js Typescript Express API",
		"contact": {
			"name": "victor.barreratoledo@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/products": {
			"get": {
				"operationId": "ListProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"products": {
											"items": {
												"$ref": "#/components/schemas/Product"
											},
											"type": "array"
										}
									},
									"required": [
										"products"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Get"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "InsertProducts",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Post",
					"Insert",
					"Create"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/Product"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}